refactor(seo): extrai transforms puros + smoke-test sem banco
- seo-transform.ts: lógica pura (slug, html, buildPostData) sem dependência
  de Prisma, permitindo teste isolado
- import-seo-programmatic.ts passa a consumir seo-transform; main() só roda
  quando executado diretamente (require.main)
- scripts/seo-import/smoke-test.ts: roda a MESMA lógica de ingestão sobre os
  primeiros N posts e valida que todas as chaves são colunas reais de BlogPost
- pnpm seo:smoke (validado: 5 posts, categorias corretas, draft/noindex, 0 chaves inválidas)
